apps/task_1/api_endpoint/User/UserRetrieve/serializer.py
from rest_framework import serializers

from apps.task_1.models import User


class UserRetrieveSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'
